requirements.txt
# EarShot Requirements
# Core dependencies for ambient speech assistant

# Audio capture
sounddevice==0.4.6
numpy

# Voice Activity Detection
onnxruntime==1.18.1

# Automatic Speech Recognition
vosk==0.3.45

# Intent Classification (sentence embeddings)
transformers==4.40.2
tokenizers>=0.19,<0.20

# LLM API client
httpx==0.27.0

# Configuration & Environment
pydantic==2.7.4
python-dotenv==1.0.1

# E-ink Display (from eink/requirements.txt)
adafruit-blinka==8.66.2
adafruit-circuitpython-epd==2.16.0
pillow==10.4.0
